moved test harness stuff into own namespace
