+++ AsyJob.IntegrationTests/Jobs/RunAJobAndPollForResponseTest.cs
@@ -1,4 +1,5 @@
-﻿using AsyJob.IntegrationTests.Users;
+﻿using AsyJob.IntegrationTests.TestHarness;
+using AsyJob.IntegrationTests.Users;
 using AsyJob.Lib.Client.Abstract.Jobs;
 using AsyJob.Lib.Jobs;
 using Microsoft.AspNetCore.Components.Forms;
