+++ AsyJob.IntegrationTests/TestHarness/ITearDown.cs
@@ -1,4 +1,4 @@
-﻿namespace AsyJob.IntegrationTests
+﻿namespace AsyJob.IntegrationTests.TestHarness
 {
     /// <summary>
     /// Responsible for running a TearDown task before an integration test is run
